requirements.txt
numpy
torch
torchvision
gym[box2d]
pybullet
gin-config
cma
opencv-python
# gym 0.20.0 reports error: https://github.com/openai/gym/issues/2432
gym[atari]==0.19.0
pyglet==1.5.27
matplotlib
moviepy
pillow
tqdm
jupyter
ipython